prevent sort reset on features list
